# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aeroguard API
  version: "1.0.0"
  description: Training simulator backend - scenario, sessions, scoring, AAR, adaptive recommendation.
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /api/v1/trainees/{trainee_id}/performance:
    get:
      summary: Historical aggregate performance for a trainee
      operationId: getPerformance
      parameters:
        - name: trainee_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Performance profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceProfile"
        "404":
          description: Unknown trainee_id
  /api/v1/training/recommend:
    post:
      summary: Recommend the next training configuration from recent performance
      operationId: recommendTraining
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendRequest"
      responses:
        "200":
          description: Recommendation with evidence-based reason
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendResponse"
        "404":
          description: Unknown trainee_id
        "422":
          description: Invalid request body
components:
  schemas:
...
    PerformanceProfile:
      type: object
      required: [trainee_id, session_count, detection_accuracy, classification_accuracy, response_accuracy]
      properties:
        trainee_id:
          type: string
        session_count:
          type: integer
        detection_accuracy:
          type: number
          minimum: 0
          maximum: 100
        classification_accuracy:
          type: number
          minimum: 0
          maximum: 100
        response_accuracy:
          type: number
          minimum: 0
          maximum: 100
        average_reaction_time_ms:
          type: number
        night_score:
          type: number
        day_score:
          type: number
        multi_threat_score:
          type: number
        low_visibility_score:
          type: number
        urban_score:
          type: number
        rural_score:
          type: number
        clear_visibility_score:
          type: number
        reduced_visibility_score:
          type: number
        poor_visibility_score:
          type: number
        high_sensor_score:
          type: number
        medium_sensor_score:
          type: number
        low_sensor_score:
          type: number
        current_level:
          type: integer
          minimum: 1
          maximum: 10
    RecommendRequest:
      type: object
      required: [trainee_id]
      properties:
        trainee_id:
          type: string
    RecommendResponse:
      type: object
      required: [recommended_difficulty, environment, time_of_day, threat_count, reason]
      properties:
        recommended_difficulty:
          type: integer
          minimum: 1
          maximum: 10
        environment:
          type: string
          enum: [urban, rural]
        time_of_day:
          type: string
          enum: [day, night]
        threat_count:
          type: integer
          enum: [1, 2, 3, 5]
        reason:
          type: string
          description: Human-readable evidence drawn from measured session metrics.